refactor files for gridworld
