+++ resources/environments/gridworld/room_environment.py
@@ -1,4 +1,3 @@
-import gym
 import numpy as np
 import torch
